merge

+++ src/main/java/com/sparta/nexusteam/vacation/service/VacationServiceImpl.java
@@ -12,8 +12,11 @@ import com.sparta.nexusteam.vacation.dto.VacationTypeResponse;
 import com.sparta.nexusteam.vacation.entity.ApprovalStatus;
 import com.sparta.nexusteam.vacation.entity.Vacation;
 import com.sparta.nexusteam.vacation.entity.VacationType;
+import com.sparta.nexusteam.vacation.entity.VacationTypeHistory;
 import com.sparta.nexusteam.vacation.repository.VacationRepository;
+import com.sparta.nexusteam.vacation.repository.VacationTypeHistoryRepository;
 import com.sparta.nexusteam.vacation.repository.VacationTypeRepository;
+import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -30,9 +33,11 @@ public class VacationServiceImpl implements VacationService {
     private final CompanyRepository companyRepository;
     private final VacationRepository vacationRepository;
     private final VacationTypeRepository vacationTypeRepository;
+    private final VacationTypeHistoryRepository vacationTypeHistoryRepository;
     private final CacheManager cacheManager;
 
     @Override
+    @Transactional
 //    @CacheEvict(value = "vacationTypes", key = "#companyId")
     public VacationTypeResponse createVacationType(PostVacationTypeRequest requestDto,
                                                    Long companyId) {
@@ -40,18 +45,29 @@ public class VacationServiceImpl implements VacationService {
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 회사가 없습니다"));
         VacationType vacationType = new VacationType(requestDto.getName(), requestDto.getDays(),
                 company);
+        VacationTypeHistory vacationTypeHistory = new VacationTypeHistory(vacationType);
         vacationType = vacationTypeRepository.save(vacationType);
+        vacationTypeHistoryRepository.save(vacationTypeHistory);
         return new VacationTypeResponse(vacationType);
     }
 
     @Override
+    @Transactional
 //    @CacheEvict(value = "vacations", key = "#employee.id")
     public VacationResponse createVacation(Long vacationTypeId, PostVacationRequest requestDto,
                                            Employee employee) {
-        VacationType vacationType = vacationTypeRepository.findById(vacationTypeId)
+        VacationTypeHistory vacationTypeHistory = vacationTypeHistoryRepository.findTopByVacationTypeIdOrderByIdDesc(
+                        vacationTypeId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 휴가 종류가 없습니다."));
+        if (requestDto.getStartDate().isAfter(requestDto.getEndDate())) {
+            throw new IllegalArgumentException("시작일시는 종료일시 보다 이르면 안됩니다.");
+        }
+        if (vacationTypeHistory.getDays() * 24 < Duration.between(requestDto.getStartDate(),
+                requestDto.getEndDate()).toHours()) {
+            throw new IllegalArgumentException(("시작일시와 종료일시의 차이는 휴가 종류 일수보다 크면 안 됩니다."));
+        }
         Vacation vacation = new Vacation(requestDto.getStartDate(), requestDto.getEndDate(),
-                vacationType, employee);
+                vacationTypeHistory, employee);
         vacation = vacationRepository.save(vacation);
         return new VacationResponse(vacation);
     }
@@ -91,7 +107,7 @@ public class VacationServiceImpl implements VacationService {
     @Override
 //    @Cacheable(value = "vacationTypes", key = "#companyId")
     public List<VacationTypeResponse> getVacationTypes(Long companyId) {
-        List<VacationType> vacationTypes = vacationTypeRepository.findByCompanyId(companyId);
+        List<VacationType> vacationTypes = vacationTypeRepository.findByCompanyIdAndIsDeletedFalse(companyId);
         return vacationTypes.stream().map(VacationTypeResponse::new).toList();
     }
 
@@ -112,6 +128,9 @@ public class VacationServiceImpl implements VacationService {
     public void deleteVacation(Long vacationId) {
         Vacation vacation = vacationRepository.findById(vacationId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 휴가는 없습니다."));
+        if (vacation.getEndDate().isAfter(LocalDateTime.now())) {
+            throw new IllegalArgumentException("사용한 휴가는 삭제 할수 없습니다.");
+        }
         vacationRepository.delete(vacation);
     }
 
@@ -122,6 +141,8 @@ public class VacationServiceImpl implements VacationService {
         VacationType vacationType = vacationTypeRepository.findById(vacationTypeId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 휴가 종류는 없습니다."));
         vacationType.updateVacationType(requestDto);
+        VacationTypeHistory vacationTypeHistory = new VacationTypeHistory(vacationType);
+        vacationTypeHistoryRepository.save(vacationTypeHistory);
         return new VacationTypeResponse(vacationType);
     }
 
@@ -130,6 +151,8 @@ public class VacationServiceImpl implements VacationService {
     public void deleteVacationType(Long vacationTypeId) {
         VacationType vacationType = vacationTypeRepository.findById(vacationTypeId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 휴가 종류는 없습니다."));
-        vacationTypeRepository.delete(vacationType);
+        vacationType.softDelete();
     }
+
+
 }
